Move NoticeTemplate to common
