+++ src/components/common/noticeTemplate/NoticeTemplate.tsx
@@ -3,7 +3,7 @@ import React, { useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Button } from 'reactstrap';
 
-import LocalizedLink from '../../common/LocalizedLink';
+import LocalizedLink from '../LocalizedLink';
 import Layout from '../../layout/Layout';
 import './noticeTemplate.scss';
