+++ src/components/common/noticeTemplate/noticeTemplate.scss
