+++ src/components/pages/notice/NotFoundPage.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 
-import NoticeTemplate from './NoticeTemplate';
+import NoticeTemplate from '../../common/noticeTemplate/NoticeTemplate';
 
 const NotFoundPage = () => {
   const { t } = useTranslation();
